move shacl tests to data model tests
