+++ backend/molgenis-emx2-datamodels/build.gradle
@@ -3,6 +3,7 @@ dependencies {
     implementation project(':backend:molgenis-emx2-io')
     implementation project(':backend:molgenis-emx2-tasks')
     implementation project(':backend:molgenis-emx2-sql')
+    testImplementation project(':backend:molgenis-emx2-rdf')
 }
 
 //we use data models from data folder
