+++ backend/molgenis-emx2-datamodels/src/test/resources/SHACL/DCAT-AP/v3.0.0/mdr-vocabularies.shape.ttl
