+++ backend/molgenis-emx2-datamodels/src/test/resources/SHACL/EJP_RD_VP_Level_1/01-08-2024/biobank.shacl
