+++ backend/molgenis-emx2-datamodels/src/test/resources/SHACL/FAIR_Data_Point/v1.2/FAIRDataPointShape.ttl
