+++ backend/molgenis-emx2-datamodels/src/test/resources/SHACL/Health-RI_core_plateau_1/v1.0.0/Resource.ttl
