+++ backend/molgenis-emx2-rdf/build.gradle
@@ -2,7 +2,7 @@ dependencies {
     implementation project(':backend:molgenis-emx2')
     implementation project(':backend:molgenis-emx2-graphql')
     testImplementation project(':backend:molgenis-emx2-sql')
-    testImplementation project(':backend:molgenis-emx2-datamodels')
     testImplementation project(':backend:molgenis-emx2-io')
     testImplementation project(':backend:molgenis-emx2-tasks')
+    testImplementation project(':backend:molgenis-emx2-datamodels')
 }
