+++ backend/molgenis-emx2-rdf/src/main/java/org/molgenis/emx2/rdf/SHACLComplianceTester.java
@@ -1,47 +1,16 @@
 package org.molgenis.emx2.rdf;
 
-import static org.molgenis.emx2.rdf.RDFTest.RDF_API_LOCATION;
-
-import java.io.ByteArrayOutputStream;
-import java.io.OutputStream;
-import org.molgenis.emx2.Database;
-import org.molgenis.emx2.Schema;
-import org.molgenis.emx2.io.ImportProfileTask;
-import org.molgenis.emx2.sql.TestDatabaseFactory;
-
-/**
- * Not a test itself, but offers to component to quickly build specific tests such as
- * CatalogueComplianceTest and FDPComplianceTest
- */
-public abstract class ComplianceTest {
-
-  /**
-   * Create a schema according to a profile and return its exported RDF
-   *
-   * @param schemaName
-   * @param profile
-   * @return
-   */
-  public static String createSchemaExportRDF(String schemaName, String profile) {
-    Database database = TestDatabaseFactory.getTestDatabase();
-    Schema schema = database.dropCreateSchema(schemaName);
-    new ImportProfileTask(schema, profile, true).run();
-    OutputStream outputStream = new ByteArrayOutputStream();
-    var rdf = new RDFService("http://localhost:8080", RDF_API_LOCATION, null);
-    rdf.describeAsRDF(outputStream, null, null, null, schema);
-    return outputStream.toString();
-  }
-
+public class SHACLComplianceTester {
   /**
    * Reusable test for compliance of RDF data to a set of SHACL files
    *
-   * @param SHACLFiles
+   * @param shaclFiles
    * @param rdf
    * @throws Exception
    */
-  public static void testCompliance(String[] SHACLFiles, String rdf) throws Exception {
+  public static void testShaclCompliance(String[] shaclFiles, String rdf) throws Exception {
     SHACLValidator sv = new SHACLValidator();
-    for (String SHACLFile : SHACLFiles) {
+    for (String SHACLFile : shaclFiles) {
       sv.addValidateShapesFromFile(SHACLFile);
     }
     sv.addValidateDataFromString(rdf);
